NU-2436 move reducers to testing
